64-bit ambiguity

+++ src/common/PJoystickHandler.cxx
@@ -736,7 +736,7 @@ VariantList PhysicalJoystickHandler::database() const
 {
   VariantList db;
   for(const auto& i: myDatabase)
-    VarList::push_back(db, i.first, i.second.joy ? i.second.joy->ID : -1);
+    VarList::push_back(db, i.first, i.second.joy ? i.second.joy->ID : Int32(-1));
 
   return db;
 }
